feat(discord-chatbot): switch to js-llmcord image
Replaces the `df-wu/llmcord` image with the `stanley2058/js-llmcord` fork to leverage its new features and active development.

A new volume for extensions has been added to support the capabilities of the `js-llmcord` image.

BREAKING CHANGE: The discord-chatbot service now uses the `js-llmcord` image. Users may need to update their `config.yaml` to be compatible with this new fork.